DRQNet/helper_requirements.txt
numpy==1.16.0
h5py==2.10.0
cython==0.29.15
open3d-python==0.3.0
pandas==0.25.3
scikit-learn==0.21.3
scipy==1.4.1
PyYAML==5.1.2